car_data/requirements.txt
django
django-chartjs
djangorestframework
psycopg2
